+++ clang/lib/Sema/SemaDeclCXX.cpp
@@ -17577,20 +17577,7 @@ static bool ConvertAPValueToString(const APValue &V, QualType T,
         }
       }
 
-      llvm::APSInt vInt = V.getInt();
-      if (llvm::APSInt::compareValues(
-              vInt, llvm::APSInt::getUnsigned(
-                        std::numeric_limits<uint64_t>::max())) >= 0 ||
-          vInt < std::numeric_limits<int64_t>::min()) {
-        // The value of cutSize is not special, it is just a number of
-        // characters that gives us enough info without losing readability.
-        const int cutSize = 20;
-        vInt.toString(Str, 16);
-        Str.erase(Str.begin() + cutSize, Str.end() - cutSize);
-        Str.insert(Str.begin() + cutSize, 3, '.');
-      } else {
-        vInt.toString(Str);
-      }
+      V.getInt().toStringTruncated(Str);
     }
 
     break;